+++ src/Angor/Avalonia/Angor.UI.Model.Implementation/Projects/StageDesign.cs
@@ -5,5 +5,5 @@ public class StageDesign : IStage
     public DateTime ReleaseDate { get; set; } = DateTime.Now;
     public long Amount { get; set; } = 12345;
     public int Index { get; set; } = 1;
-    public double RatioOfTotal { get; set; } = 1;
+    public decimal RatioOfTotal { get; set; } = 1m;
 }
