+++ src/Angor/Avalonia/AngorApp/Features/Invest/Amount/Breakdown.cs
@@ -1,6 +1,6 @@
 namespace AngorApp.Features.Invest.Amount;
 
-public record Breakdown(int Index, IAmountUI InvestAmount, double RatioOfTotal, DateTimeOffset ReleaseDate)
+public record Breakdown(int Index, IAmountUI InvestAmount, decimal RatioOfTotal, DateTimeOffset ReleaseDate)
 {
     public IAmountUI Amount => new AmountUI((long)(InvestAmount.Sats * RatioOfTotal));
 }
